requirements.txt
opencv-python
pytesseract
openai
python-dotenv